requirements.txt
flask==3.0.3
torch>=2.1.0
transformers>=4.41.0
numpy>=1.24.0
python-docx
PyPDF2
